msm: kgsl: Fix command completion race with profile IB write
Ensure that the submit data for user profiling is written
out to the profiling IB before we submit work to the GMU
dispatch queue. Otherwise, if the GMU is ready to dispatch
and the GPU work is small, it can so happen that the command
retires even before the user profiling data gets written out
completely.

Change-Id: I8df39666366e8579acc9151988881ecef3c8f1a4
Signed-off-by: Sushmita Susheelendra <quic_ssusheel@quicinc.com>
Signed-off-by: Deepak Kumar <quic_dkumar@quicinc.com>
Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>