techpack: touch: focaltech_3680: Remove obsolete charging status check
fts_get_charging_status() invokes power_supply_is_system_supplied() to check
if the device is charging. But we can already do that by just checking
POWER_SUPPLY_PROP_ONLINE which is already done in fts_get_charging_status()
later on. power_supply_is_system_supplied() is meaningless here. And it also
passes the USB SCOPE property to the charger driver that causes a logspam of
unavailable property in the charger driver, which makes it all the more
pointless.

Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>